Tested get and updated state info
